Don't treat thirdlines from nick 'Mode' as Mode changes in weechat3
If anyone has the nick 'Mode' and appears in a thirdline, this line will be
treated as a Mode change of another nick, and will fail to be parsed.
This happens when, for example, a user changes their nick from 'Mode' to
something else.

This patch changes the order the different thirdline types are detected so that
a mode change is the very last type. This means that nick changes and other
thirdlines are caught first and handled correctly.